Add classes to work with matrixes

+++ src/com/example/cs_c_matrix_calc/apapter/MatrixGridAdapter.java
@@ -3,29 +3,26 @@ package com.example.cs_c_matrix_calc.apapter;
 import android.content.Context;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.EditText;
-import android.widget.TextView;
-
 import com.example.cs_c_matrix_calc.R;
 
 /**
  * Created by cs_c on 5/15/14.
  */
-public class MatrixGridAdapter extends BaseAdapter{
+public class MatrixGridAdapter extends BaseAdapter {
     private int mMatrixSize;
     private Context mContext;
-    public int[][] mNumberArray;
+    public double[][] mNumberArray;
 
 
     public MatrixGridAdapter(Context _context, int _size) {
         mContext = _context;
         mMatrixSize = _size;
-        mNumberArray = new int[_size][_size];
+        mNumberArray = new double[_size][_size];
 
     }
 
@@ -60,7 +57,7 @@ public class MatrixGridAdapter extends BaseAdapter{
             holder = (ViewHolder) convertView.getTag();
         }
         holder.etComment.setTag(position);
-        holder.etComment.setHint((position/mMatrixSize)+1 + "x" + ((position%mMatrixSize)+1));
+        holder.etComment.setHint((position / mMatrixSize) + 1 + "x" + ((position % mMatrixSize) + 1));
         final ViewHolder finalHolder = holder;
         holder.etComment.addTextChangedListener(new TextWatcher() {
 
@@ -74,7 +71,7 @@ public class MatrixGridAdapter extends BaseAdapter{
             public void onTextChanged(CharSequence s, int start,
                                       int before, int count) {
                 int pos = Integer.parseInt(finalHolder.etComment.getTag().toString());
-                mNumberArray[pos/mMatrixSize][pos%mMatrixSize] =
+                mNumberArray[pos / mMatrixSize][pos % mMatrixSize] =
                         Integer.parseInt(finalHolder.etComment.getText().toString());
             }
         });
